cool

+++ commands/say.js
@@ -3,7 +3,8 @@ module.exports = {
     description: "Say anything",
     execute(message, args) {
         const sayMessage = message.content.split(" ").slice(1).join(" ");
+        if (!sayMessage) return;
         message.channel.send(sayMessage);
         message.delete();
-    }
-}
+    },
+};
